chore: cache pools info
