+++ packages/oraidex-server/package.staging.json
@@ -1,6 +1,6 @@
 {
   "name": "@oraichain/oraidex-server-staging",
-  "version": "1.0.51",
+  "version": "1.0.52",
   "main": "dist/index.js",
   "bin": "dist/index.js",
   "license": "MIT",
