+++ packages/oraidex-server/src/helper.ts
@@ -1,18 +1,27 @@
-import { fetchRetry, ROUTER_V2_CONTRACT } from "@oraichain/oraidex-common";
-import bech32 from "bech32";
 import { CosmWasmClient } from "@cosmjs/cosmwasm-stargate";
+import { fetchRetry, ROUTER_V2_CONTRACT } from "@oraichain/oraidex-common";
 import { AssetInfo, OraiswapRouterQueryClient } from "@oraichain/oraidex-contracts-sdk";
 import {
-  simulateSwapPrice,
+  injAddress,
+  ORAI,
+  oraiInfo,
+  oraixCw20Address,
+  PairMapping,
   pairs,
   parseAssetInfoOnlyDenom,
-  ORAI,
-  injAddress,
-  oraiInfo,
-  PairMapping,
-  oraixCw20Address,
-  usdcCw20Address
+  simulateSwapPrice,
+  usdcCw20Address,
+  getAllFees,
+  getAllVolume24h,
+  getPoolAmounts,
+  getPoolLiquidities,
+  PairInfoDataResponse,
+  getPoolsFromDuckDb,
+  getPoolAprsFromDuckDb
 } from "@oraichain/oraidex-sync";
+import bech32 from "bech32";
+import "dotenv/config";
+
 const rpcUrl = process.env.RPC_URL || "https://rpc.orai.io";
 
 export function parseSymbolsToTickerId([base, quote]: [string, string]) {
@@ -119,4 +128,53 @@ export const fetchSimulatePrices = async () => {
   );
 
   return prices;
+};
+
+export const getAllPoolsInfo = async () => {
+  try {
+    const volumes = await getAllVolume24h();
+    const allFee7Days = await getAllFees();
+    const pools = await getPoolsFromDuckDb();
+    const allPoolApr = await getPoolAprsFromDuckDb();
+    const allLiquidities = await getPoolLiquidities(pools);
+    const allPoolAmounts = await getPoolAmounts(pools);
+
+    const allPoolsInfo: PairInfoDataResponse[] = pools.map((pool, index) => {
+      const poolApr = allPoolApr.find((item) => item.pairAddr === pool.pairAddr);
+      if (!poolApr) return null;
+
+      const poolFee = allFee7Days.find((item) => {
+        const [baseAssetInfo, quoteAssetInfo] = item.assetInfos;
+        return (
+          JSON.stringify(baseAssetInfo) === pool.firstAssetInfo &&
+          JSON.stringify(quoteAssetInfo) === pool.secondAssetInfo
+        );
+      });
+
+      const poolVolume = volumes.find((item) => {
+        const [baseAssetInfo, quoteAssetInfo] = item.assetInfos;
+        return (
+          JSON.stringify(baseAssetInfo) === pool.firstAssetInfo &&
+          JSON.stringify(quoteAssetInfo) === pool.secondAssetInfo
+        );
+      });
+      if (!poolVolume) return null;
+
+      return {
+        ...pool,
+        volume24Hour: poolVolume.volume.toString(),
+        fee7Days: poolFee.fee.toString(),
+        apr: poolApr.apr,
+        totalLiquidity: allLiquidities[index],
+        rewardPerSec: poolApr.rewardPerSec,
+        offerPoolAmount: allPoolAmounts[index].offerPoolAmount,
+        askPoolAmount: allPoolAmounts[index].askPoolAmount,
+        totalSupply: poolApr.totalSupply
+      } as PairInfoDataResponse;
+    });
+
+    return allPoolsInfo;
+  } catch (error) {
+    console.log({ errorGetAllPoolsInfo: error });
+  }
 };
