+++ packages/oraidex-server/src/map-cache.ts
@@ -1,5 +1,6 @@
 export enum CACHE_KEY {
-  SIMULATE_PRICE = "SIMULATE_PRICE"
+  SIMULATE_PRICE = "SIMULATE_PRICE",
+  POOLS_INFO = "POOLS_INFO"
 }
 export const cache: Map<string, any> = new Map();
 export const cacheListeners: Map<string, (...args: any[]) => Promise<void>> = new Map();
@@ -10,6 +11,7 @@ export const registerListener = (key: string, listener: (...args: any[]) => Prom
 
 export const updateInterval = async (interval = 5000): Promise<void> => {
   for (const key of cacheListeners.keys()) {
+    console.log("rerun");
     const listener = cacheListeners.get(key);
     if (listener) {
       const value = await listener();
