+++ packages/oraidex-sync/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@oraichain/oraidex-sync",
-  "version": "1.0.41",
+  "version": "1.0.0",
   "main": "build/index.js",
   "license": "MIT",
   "files": [
